Make the Action Bar Visible
This commit makes the action bar finally visible by fixing the build.gradle file on both the Stream3D app itself and the Gaminganywhere module. Specifically, gaminganywhere module now extends the AppCompatActivity instead of just activity. To achieve that, we need also to standarize the compileSdkVersion, minSdkVersion, and targetSdkVersion between the build.gradle of Stream3D and Gaminganywhere.